coding_agent/requirements.txt
# Core dependencies for OverwatchAgent
google-adk>=0.1.0
pydantic>=2.0.0

# Voice input dependencies
SpeechRecognition>=3.10.0
PyAudio>=0.2.11

# Optional dependencies for enhanced functionality
python-dotenv>=1.0.0  # For .env file support
